separated player from square
